+++ src/embedded/sink.rs
@@ -11,13 +11,14 @@ use serde::Serialize;
 use crate::embedded;
 use crate::embedded::Env;
 use crate::errors::*;
+use crate::event;
 use crate::event::Event;
 use crate::json_schema;
 use crate::lib::Config;
 
 pub fn read_hashes(file: File) -> Result<HashSet<String>> {
     let reader = BufReader::new(file);
-    let events = embedded::events(reader);
+    let events = event::events(reader);
     let mut hashes = HashSet::new();
 
     for result in events {