+++ src/event.rs
@@ -1,4 +1,5 @@
 use std::collections::HashMap;
+use std::io::{BufRead, BufReader, Read};
 
 use multihash::MultihashDigest;
 use serde::Deserialize;
@@ -24,4 +25,14 @@ pub fn event_hash(mut event: Event) -> Result<String> {
     let hash = multihash::Code::Sha2_256.digest(&bytes);
     let base58 = bs58::encode(hash.to_bytes());
     Ok(base58.into_string())
+}
+
+pub fn parse_event(s: &str) -> Result<Event> {
+    serde_json::from_str(s).chain_err(|| "Event deserialization failed")
+}
+
+pub fn events(reader: BufReader<impl Read>) -> impl Iterator<Item = Result<Event>> {
+    reader
+        .lines()
+        .map(|line| parse_event(line.chain_err(|| "Failed to read line")?.as_str()))
 }